requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
langchain==0.0.325
langchain-community
langdetect==1.0.9
chromadb==0.4.15
python-dotenv==1.0.0
python-docx==0.8.11
cohere==4.32
